[libc++] Split the multidimensional algorithms out of uninitialized_algorithms.h
